+++ src/common-lib/urls/urls.ts
@@ -118,13 +118,19 @@ export type LessonListingLinkProps = {
   programmeSlug: string;
   unitSlug: string;
 };
-export type UnitPageLinkProps = {
-  page: "unit-page";
-  subjectPhaseSlug: string;
+export type IntegratedLessonListingLinkProps = {
+  page: "integrated-lesson-index";
+  programmeSlug: string;
   unitSlug: string;
   query?: {
     subject_category?: string;
   };
+};
+export type IntegratedLessonOverviewLinkProps = {
+  page: "integrated-lesson-overview";
+  programmeSlug: string;
+  unitSlug: string;
+  lessonSlug: string;
 };
 export type SpecialistLessonListingLinkProps = Omit<
   LessonListingLinkProps,
@@ -451,7 +457,8 @@ export type OakLinkProps =
   | SpecialistLessonOverviewLinkProps
   | LessonOverviewCanonicalLinkProps
   | LessonListingLinkProps
-  | UnitPageLinkProps
+  | IntegratedLessonListingLinkProps
+  | IntegratedLessonOverviewLinkProps
   | SpecialistLessonListingLinkProps
   | UnitListingLinkProps
   | SpecialistUnitListingLinkProps
@@ -834,11 +841,18 @@ export const OAK_PAGES: {
     configType: "internal",
     pageType: "lesson-index",
   }),
-  "unit-page": createOakPageConfig({
-    pathPattern: "/programmes/:subjectPhaseSlug/units/:unitSlug/lessons",
+  "integrated-lesson-index": createOakPageConfig({
+    pathPattern: "/programmes/:programmeSlug/units/:unitSlug/lessons",
     analyticsPageName: "Lesson Listing",
     configType: "internal",
-    pageType: "unit-page",
+    pageType: "integrated-lesson-index",
+  }),
+  "integrated-lesson-overview": createOakPageConfig({
+    pathPattern:
+      "/programmes/:programmeSlug/units/:unitSlug/lessons/:lessonSlug",
+    analyticsPageName: "Lesson",
+    configType: "internal",
+    pageType: "integrated-lesson-overview",
   }),
   "specialist-lesson-index": createOakPageConfig({
     pathPattern: